+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
+## [v1.2.0] - 2021-02-10
+- Added Messenger Platform types
+- Changed node info read type key from `type` to `read_return` (breaking change in certain cases)
+- Added delete method on nodes
+## [v1.1.0] - 2021-02-08
+- Added create method on edges
 ## [v1.0.1] - 2021-02-03
 - Initial release (republished)